+++ internal/pkg/runtime/engine/apptainer/prepare_linux.go
@@ -3,7 +3,7 @@
 //   For website terms of use, trademark policy, privacy policy and other
 //   project policies see https://lfprojects.org/policies
 // Copyright (c) 2020, Control Command Inc. All rights reserved.
-// Copyright (c) 2018-2023, Sylabs Inc. All rights reserved.
+// Copyright (c) 2018-2025, Sylabs Inc. All rights reserved.
 // This software is licensed under a 3-clause BSD license. Please consult the
 // LICENSE.md file distributed with the sources of this project regarding your
 // rights to use or distribute this software.
@@ -25,10 +25,10 @@ import (
 	"github.com/ProtonMail/go-crypto/openpgp"
 	"github.com/apptainer/apptainer/internal/pkg/buildcfg"
 	"github.com/apptainer/apptainer/internal/pkg/cgroups"
+	"github.com/apptainer/apptainer/internal/pkg/fakeroot"
 	fakerootutil "github.com/apptainer/apptainer/internal/pkg/fakeroot"
 	"github.com/apptainer/apptainer/internal/pkg/image/driver"
 	"github.com/apptainer/apptainer/internal/pkg/instance"
-	"github.com/apptainer/apptainer/internal/pkg/plugin"
 	"github.com/apptainer/apptainer/internal/pkg/runtime/engine/config/starter"
 	"github.com/apptainer/apptainer/internal/pkg/security"
 	"github.com/apptainer/apptainer/internal/pkg/security/seccomp"
@@ -41,7 +41,6 @@ import (
 	"github.com/apptainer/apptainer/internal/pkg/util/mainthread"
 	"github.com/apptainer/apptainer/internal/pkg/util/user"
 	"github.com/apptainer/apptainer/pkg/image"
-	fakerootcallback "github.com/apptainer/apptainer/pkg/plugin/callback/runtime/fakeroot"
 	apptainerConfig "github.com/apptainer/apptainer/pkg/runtime/engine/apptainer/config"
 	"github.com/apptainer/apptainer/pkg/runtime/engine/config"
 	"github.com/apptainer/apptainer/pkg/sylog"
@@ -678,21 +677,8 @@ func (e *EngineOperations) prepareContainerConfig(starterConfig *starter.Config)
 			}
 		}
 
-		getIDRange := fakerootutil.GetIDRange
-
-		callbackType := (fakerootcallback.UserMapping)(nil)
-		callbacks, err := plugin.LoadCallbacks(callbackType)
-		if err != nil {
-			return fmt.Errorf("while loading plugins callbacks '%T': %s", callbackType, err)
-		}
-		if len(callbacks) > 1 {
-			return fmt.Errorf("multiple plugins have registered hook callback for fakeroot")
-		} else if len(callbacks) == 1 {
-			getIDRange = callbacks[0].(fakerootcallback.UserMapping)
-		}
-
 		e.EngineConfig.OciConfig.AddLinuxUIDMapping(uid, 0, 1)
-		idRange, err := getIDRange(fakerootutil.SubUIDFile, uid)
+		idRange, err := fakeroot.GetUIDRange(uid)
 		if err != nil {
 			return fmt.Errorf("could not use fakeroot: %s", err)
 		}
@@ -700,7 +686,7 @@ func (e *EngineOperations) prepareContainerConfig(starterConfig *starter.Config)
 		starterConfig.AddUIDMappings(e.EngineConfig.OciConfig.Linux.UIDMappings)
 
 		e.EngineConfig.OciConfig.AddLinuxGIDMapping(gid, 0, 1)
-		idRange, err = getIDRange(fakerootutil.SubGIDFile, uid)
+		idRange, err = fakeroot.GetGIDRange(uid)
 		if err != nil {
 			return fmt.Errorf("could not use fakeroot: %s", err)
 		}
